+++ lib/rubocop-graphql.rb
@@ -7,8 +7,12 @@ require_relative "rubocop/graphql/version"
 require_relative "rubocop/graphql/inject"
 require_relative "rubocop/graphql/node_pattern"
 
+require_relative "rubocop/graphql/argument"
+require_relative "rubocop/graphql/argument/block"
+require_relative "rubocop/graphql/argument/kwargs"
 require_relative "rubocop/graphql/field"
-require_relative "rubocop/graphql/field_kwargs"
+require_relative "rubocop/graphql/field/block"
+require_relative "rubocop/graphql/field/kwargs"
 require_relative "rubocop/graphql/schema_member"
 
 RuboCop::GraphQL::Inject.defaults!